+++ src/loss.hpp
@@ -8,54 +8,54 @@ namespace ogbt {
 
 template<typename Score, typename Residual> class Loss {
 public:
-  static double score(const std::vector<double> &y_pred, const std::vector<double> &y_true) noexcept {
+  static float score(const std::vector<float> &y_pred, const std::vector<float> &y_true) noexcept {
     return Score::evaluate(y_pred, y_true);
   }
 
-  static double score(const std::vector<double> &y_pred, const Dataset &data) noexcept {
+  static float score(const std::vector<float> &y_pred, const Dataset &data) noexcept {
     return score(y_pred, data.get_y());
   }
 
   template<typename TLoss, typename TreeGenAlgo>
-  static double score(const Model<TLoss> &model, const Dataset &data) noexcept {
+  static float score(const Model<TLoss> &model, const Dataset &data) noexcept {
     return score(model.predict(data.get_x()), data.get_y());
   }
 
-  static double score(const Tree &tree, const Dataset &data) noexcept {
+  static float score(const Tree &tree, const Dataset &data) noexcept {
     return score(tree.predict(data.get_x()), data);
   }
 
-  static std::vector<double> residual(const std::vector<double> &y_pred, const std::vector<double> &y_true) noexcept {
+  static std::vector<float> residual(const std::vector<float> &y_pred, const std::vector<float> &y_true) noexcept {
     return Residual::evaluate(y_pred, y_true);
   }
 
-  static std::vector<double> residual(const std::vector<double> &y_pred, const Dataset &data) noexcept {
+  static std::vector<float> residual(const std::vector<float> &y_pred, const Dataset &data) noexcept {
     return residual(y_pred, data.get_y());
   }
 
   template<typename TLoss, typename TreeGenAlgo>
-  static std::vector<double> residual(const Model<TLoss> &model, Dataset data) noexcept {
+  static std::vector<float> residual(const Model<TLoss> &model, Dataset data) noexcept {
     return residual(model.predict(data.get_x()), data);
   }
 
-  static std::vector<double> residual(const Tree &tree, Dataset data) noexcept {
+  static std::vector<float> residual(const Tree &tree, Dataset data) noexcept {
     return residual(tree.predict(data.get_x()), data);
   }
 };
 
 
 struct ScoreMSE {
-  static double evaluate(const std::vector<double> &y_pred, const std::vector<double> &y_true) noexcept {
+  static float evaluate(const std::vector<float> &y_pred, const std::vector<float> &y_true) noexcept {
     assert(y_pred.size() == y_true.size());
-    double result = 0;
+    float result = 0;
     for (size_t i = 0; i < y_pred.size(); i++) { result += (y_true[i] - y_pred[i]) * (y_true[i] - y_pred[i]); }
     return result / y_true.size();
   }
 };
 
 struct ResidualMSE {
-  static std::vector<double> evaluate(const std::vector<double> &y_pred, const std::vector<double> &y_true) noexcept {
-    std::vector<double> ans;
+  static std::vector<float> evaluate(const std::vector<float> &y_pred, const std::vector<float> &y_true) noexcept {
+    std::vector<float> ans;
     ans.resize(y_true.size());
     for (size_t i = 0; i < y_true.size(); i++) { ans[i] = y_true[i] - y_pred[i]; }
     return ans;
@@ -63,9 +63,9 @@ struct ResidualMSE {
 };
 
 struct ScoreLogLoss {
-  static double evaluate(const std::vector<double> &y_pred, const std::vector<double> &y_true) noexcept {
+  static float evaluate(const std::vector<float> &y_pred, const std::vector<float> &y_true) noexcept {
     assert(y_pred.size() == y_true.size());
-    double result = 0;
+    float result = 0;
     for (size_t i = 0; i < y_pred.size(); i++) {
       auto sigmoid_val = 1.0 / (1.0 + exp(-y_pred[i]));
       result += y_true[i] * log(sigmoid_val) + (1 - y_true[i]) * log(1 - sigmoid_val);
@@ -75,8 +75,8 @@ struct ScoreLogLoss {
 };
 
 struct ResidualLogLoss {
-  static std::vector<double> evaluate(const std::vector<double> &y_pred, const std::vector<double> &y_true) noexcept {
-    std::vector<double> ans;
+  static std::vector<float> evaluate(const std::vector<float> &y_pred, const std::vector<float> &y_true) noexcept {
+    std::vector<float> ans;
     ans.resize(y_true.size());
     for (size_t i = 0; i < y_true.size(); i++) {
       auto sigmoid_val = 1.0 / (1.0 + exp(-y_pred[i]));
