+++ tests/tree_test.cpp
@@ -3,8 +3,8 @@
 
 TEST(Tree, Constructor) {
   std::vector<std::vector<int>> data_categorical = { { 0, 1 }, { 1, 1 } };
-  std::vector<std::vector<double>> data_dense = { { 1.2, -10.0 } };
-  std::vector<double> target = { 1.0, 0.0 };
+  std::vector<std::vector<float>> data_dense = { { 1.2, -10.0 } };
+  std::vector<float> target = { 1.0, 0.0 };
   ogbt::Dataset dataset(data_categorical, data_dense, target);
   std::mt19937 gen;
   ogbt::Tree tree(dataset, gen, 6);
@@ -21,8 +21,8 @@ TEST(Tree, Constructor) {
 
 TEST(Tree, Predict) {
   std::vector<std::vector<int>> data_categorical = { { 0, 1 }, { 1, 1 } };
-  std::vector<std::vector<double>> data_dense = { { 1.2, -10.0 } };
-  std::vector<double> target = { 1.0, 0.0 };
+  std::vector<std::vector<float>> data_dense = { { 1.2, -10.0 } };
+  std::vector<float> target = { 1.0, 0.0 };
   ogbt::Dataset dataset(data_categorical, data_dense, target);
   std::mt19937 gen;
   ogbt::Tree tree(dataset, gen);
